+++ api/CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 #### Added
 
-* Added support for S3 as a storage backend (#[11](https://github.com/stjude-rust-labs/planetary/pull/11)).
+* Added support for Google Cloud Storage as a storage backend ([#12](https://github.com/stjude-rust-labs/planetary/pull/12)).
+* Added support for S3 as a storage backend ([#11](https://github.com/stjude-rust-labs/planetary/pull/11)).
 * Added `--transporter-image` CLI option to specify the `planetary-transporter`
   image to use for inputs and outputs pods ([#8](https://github.com/stjude-rust-labs/planetary/pull/8)).
 * Initial implementation of the full TES API and database abstraction ([#4](https://github.com/stjude-rust-labs/planetary/pull/4)).